Add install command to briefcase
