+++ src/briefcase/commands/__init__.py
@@ -1,6 +1,7 @@
 from .build import BuildCommand  # noqa
 from .create import CreateCommand  # noqa
 from .dev import DevCommand  # noqa
+from .install import InstallCommand  # noqa
 from .new import NewCommand  # noqa
 from .package import PackageCommand  # noqa
 from .publish import PublishCommand  # noqa
